+++ jest.config.cjs
@@ -62,7 +62,10 @@ module.exports = {
 	// A set of global variables that need to be available in all test environments
 	globals: {
 		"ts-jest": {
-			"useESM": true
+			"useESM": true,
+			diagnostics: {
+				exclude: ['!test/**/*.ts'],
+			},
 		}
 	},